backend/requirements.txt
numpy==1.23.5
openai==0.27.2
pandas==1.5.3
pydub==0.25.1
pygame==2.1.3.dev8
pyttsx3==2.90
requests==2.28.2
Flask==2.1.1
gunicorn==20.1.0
python-dotenv